Rename GF.Compile.Compute.ConcreteNew to GF.Compile.Compute.Concrete
